+++ CHANGELOG.rst
@@ -10,6 +10,7 @@ Changelog
 * Improve CNV caller for CYP2A6, CYP2D6, and SULT1A1.
 * Add new method :meth:`sdk.utils.add_cn_samples` method.
 * Update :command:`compare-genotypes` command to output CNV comparison results as well.
+* Update :command:`estimate-phase-beagle` command. From now on, the 'chr' prefix in contig names (e.g. 'chr1' vs. '1') will be automatically added or removed as necessary to match the reference VCF’s contig names.
 
 0.10.1 (2021-12-20)
 -------------------